add handling for nullable variables

+++ benchmon/benchmark/constraints/resctrl.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Iterable, TYPE_CHECKING, Tuple
+from typing import Iterable, Optional, TYPE_CHECKING, Tuple
 
 from .base import BaseConstraint
 from .. import BaseBenchmark
@@ -18,7 +18,7 @@ class ResCtrlConstraint(BaseConstraint):
     resctrl를 통해 입력받은 값으로 제한하며, 벤치마크의 실행이 종료될 경우 그 resctrl 그룹을 삭제한다.
     """
     _masks: Tuple[str, ...]
-    _group: ResCtrl
+    _group: Optional[ResCtrl] = None
 
     def __init__(self, masks: Iterable[str]) -> None:
         """
@@ -26,13 +26,11 @@ class ResCtrlConstraint(BaseConstraint):
         :type masks: typing.Iterable[str]
         """
         self._masks = tuple(masks)
-        self._group = ResCtrl()
 
     async def on_start(self, context: Context) -> None:
         benchmark = BaseBenchmark.of(context)
 
-        self._group.group_name = benchmark.group_name
-
+        self._group = ResCtrl(benchmark.group_name)
         self._group.create_group()
 
         if len(self._masks) is not 0:
@@ -42,7 +40,5 @@ class ResCtrlConstraint(BaseConstraint):
         await self._group.add_tasks(children)
 
     async def on_destroy(self, context: Context) -> None:
-        try:
+        if self._group is not None:
             await self._group.delete()
-        except PermissionError:
-            pass
